fix type mismatches

+++ src/nodes/reducemean.cc
@@ -20,9 +20,9 @@ void ReduceMean::parseAttributes( onnx::NodeProto &node )
 	Tensor* input = get_input_tensor(0);
 
 	// Default to reducing all dimensions
-	for (int i = 0; i < input->data_dim.size(); ++i)
+	for (size_t i = 0; i < input->data_dim.size(); ++i)
 	{
-		axes.push_back(i);
+		axes.push_back(static_cast<int64_t>(i));
 	}
 
 	// Default to keeping all dimensions
@@ -38,7 +38,7 @@ void ReduceMean::parseAttributes( onnx::NodeProto &node )
 			{
 				while (axis < 0)
 				{
-					axis += get_input_tensor(0)->data_dim.size();
+					axis += static_cast<int64_t>(get_input_tensor(0)->data_dim.size());
 				}
 			}
 		}
@@ -60,9 +60,9 @@ void ReduceMean::resolve(void)
 	Tensor *t = new Tensor;
 	t->data_dim = {};
 
-	for (int i = 0; i < input->data_dim.size(); ++i)
+	for (size_t i = 0; i < input->data_dim.size(); ++i)
 	{
-		if(std::find(axes.begin(), axes.end(), i) != axes.end())
+		if(std::find(axes.begin(), axes.end(), static_cast<int64_t>(i)) != axes.end())
 		{
 			if (keepdims)
 			{
@@ -91,12 +91,12 @@ void ReduceMean::print(std::ostream &dst) const
 	// Allocate arrays for intermediate sums/means (final reduction goes directly to output)
 	std::vector<int> sumDims = input->data_dim; // used to track axis reductions based on the keepdims attribute
 	std::unordered_map<int64_t, std::vector<int>> axisDims; // dimensions at each reduction
-	for (int i = 0; i < axes.size() - 1; ++i)
+	for (size_t i = 0; i < axes.size() - 1; ++i)
 	{
 		int64_t axis = axes[i];
 		INDT_1 << input->data_type_str() << " sum_axis_" << axis;
 		sumDims[axis] = keepdims ? 1 : 0;
-		for (int i = 0; i < sumDims.size(); ++i)
+		for (size_t i = 0; i < sumDims.size(); ++i)
 		{
 			if (sumDims[i])
 			{
@@ -114,7 +114,7 @@ void ReduceMean::print(std::ostream &dst) const
 	for (int64_t axis : axes)
 	{
 		INDT_1 << "int dims_" << axis << "[" << axisDims[axis].size() << "] = {";
-		for (int i = 0; i < axisDims[axis].size(); ++i)
+		for (size_t i = 0; i < axisDims[axis].size(); ++i)
 		{
 			dst << axisDims[axis][i];
 			if (i < axisDims[axis].size() - 1)
@@ -138,7 +138,7 @@ void ReduceMean::print(std::ostream &dst) const
 
 	// Set to 0
 	INDT_2 << kOutputName;
-	for (int i = 0; i < output->data_dim.size(); ++i)
+	for (size_t i = 0; i < output->data_dim.size(); ++i)
 	{
 		dst << "[loc[" << i << "]]";
 	}
@@ -148,7 +148,7 @@ void ReduceMean::print(std::ostream &dst) const
 
 	// Reduce each requested axis
 	std::string prevTensorName = kInputName; // remember previous tensor to use for next sum
-	for (int i = 0; i < axes.size(); ++i)
+	for (size_t i = 0; i < axes.size(); ++i)
 	{
 		// Comment the axis being reduced
 		INDT_1 << "// Reduce axis " << axes[i] << std::endl;
@@ -171,14 +171,14 @@ void ReduceMean::print(std::ostream &dst) const
 
 		// Add to sum
 		INDT_3 << outputTensor;
-		for (int i = 0; i < currentAxisDims.size(); ++i)
+		for (size_t i = 0; i < currentAxisDims.size(); ++i)
 		{
 			dst << "[loc[" << i << "]]";
 		}
 		dst << " += " << prevTensorName;
 		// Get dimensions of tensor before reduction
 		const std::vector<int>& beforeReductionDims = (i == 0) ? input->data_dim : axisDims[axes[i - 1]];
-		for (int i = 0, locIndex = 0; i < beforeReductionDims.size(); ++i)
+		for (size_t i = 0, locIndex = 0; i < beforeReductionDims.size(); ++i)
 		{
 			if (i == axis)
 			{
@@ -208,11 +208,11 @@ void ReduceMean::print(std::ostream &dst) const
 		INDT_1 << "for (int j = 0; j < " << totalElements << "; ++j) {" << std::endl;
 
 		// Tracking sum index
-		printLocationArray(dst, 2, axis, currentAxisDims.size(), "j");
+		printLocationArray(dst, 2, axis, static_cast<int>(currentAxisDims.size()), "j");
 
 		// Convert sum to mean
 		INDT_2 << outputTensor;
-		for (int i = 0; i < currentAxisDims.size(); ++i)
+		for (size_t i = 0; i < currentAxisDims.size(); ++i)
 		{
 			dst << "[loc[" << i << "]]";
 		}
